requirements.txt
langchain-openai
langchain-anthropic
langchain-mistralai
langchain-community 
langchain-text-splitters 
langchain-core 
langchain-huggingface 
langgraph 
langgraph-checkpoint 
huggingface_hub 
pymupdf
dotenv